+++ CHANGELOG.md
@@ -6,10 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.2] - 2022-05-24
+### Changed
+- InsufficientNearDeposit -> InsufficientDeposit
 
-## [2.0.1] - 2022-05-22
+## [2.0.1] - 2022-05-20
 ### Added
-- View-method get_all_streams and test for it
+- Transformation of wrapped NEAR into unwrapped ones at Finance contract
+### Changed
+- Gas needs for withdrawal were increased
 
 ## [2.0.0] - 2022-05-16
 ### Added